+++ packages/paste-core/components/topbar/src/Topbar.tsx
@@ -27,13 +27,12 @@ export const Topbar = React.forwardRef<HTMLDivElement, TopbarProps>(
         borderBottomStyle="solid"
         borderBottomColor="colorBorderWeaker"
         paddingX="space70"
-        paddingY="space60"
         position="sticky"
         top="0"
         display="flex"
         justifyContent="space-between"
         alignItems="center"
-        zIndex="zIndex10"
+        zIndex="zIndex40"
       >
         {children}
       </Box>