+++ packages/paste-website/src/components/ContactUsMenu.tsx
@@ -1,35 +1,11 @@
 import * as React from 'react';
-import {Box} from '@twilio-paste/box';
-import {Text} from '@twilio-paste/text';
 import {Menu, MenuButton, MenuItem, useMenuState} from '@twilio-paste/menu';
-import {MediaObject, MediaFigure, MediaBody} from '@twilio-paste/media-object';
-import {MoreIcon} from '@twilio-paste/icons/esm/MoreIcon';
 import {ProductSupportIcon} from '@twilio-paste/icons/esm/ProductSupportIcon';
 
 import {event} from '../lib/gtag';
 
-interface ContactUsMenuProps {
-  placement?:
-    | 'auto-start'
-    | 'auto'
-    | 'auto-end'
-    | 'top-start'
-    | 'top'
-    | 'top-end'
-    | 'right-start'
-    | 'right'
-    | 'right-end'
-    | 'bottom-end'
-    | 'bottom'
-    | 'bottom-start'
-    | 'left-end'
-    | 'left'
-    | 'left-start'
-    | undefined;
-}
-
-export const ContactUsMenu: React.FC<React.PropsWithChildren<ContactUsMenuProps>> = ({placement = 'right-end'}) => {
-  const menu = useMenuState({placement});
+export const ContactUsMenu: React.FC = () => {
+  const menu = useMenuState();
 
   const handleClick = (category: string, action: string, label: string): void => {
     menu.hide();
@@ -41,35 +17,20 @@ export const ContactUsMenu: React.FC<React.PropsWithChildren<ContactUsMenuProps>
   };
 
   return (
-    <Box marginTop="auto" padding="space20" paddingBottom={['space0', 'space0', 'space70']}>
+    <>
       <MenuButton
         {...menu}
-        variant="reset"
+        variant="secondary_icon"
         size="reset"
-        fullWidth
         onClick={() =>
           event({
-            category: 'Left Navigation',
+            category: 'Top Navigation',
             action: 'click-contact-us',
             label: 'Contact us',
           })
         }
       >
-        <Box as="span" display="flex" alignItems="center" padding="space40" width="100%">
-          <MediaObject verticalAlign="center">
-            <MediaFigure spacing="space30">
-              <ProductSupportIcon decorative />
-            </MediaFigure>
-            <MediaBody>
-              <Text as="span" fontSize={['fontSize50', 'fontSize50', 'fontSize30']} lineHeight="lineHeight20">
-                Contact us
-              </Text>
-            </MediaBody>
-          </MediaObject>
-          <Box as="span" marginLeft="auto">
-            <MoreIcon decorative />
-          </Box>
-        </Box>
+        <ProductSupportIcon decorative={false} title="Contact us" />
       </MenuButton>
       <Menu {...menu} aria-label="Contact us">
         <MenuItem
@@ -94,6 +55,6 @@ export const ContactUsMenu: React.FC<React.PropsWithChildren<ContactUsMenuProps>
           Report a bug
         </MenuItem>
       </Menu>
-    </Box>
+    </>
   );
 };