+++ packages/paste-website/src/components/tokens-list/TokensListFilter.tsx
@@ -9,6 +9,8 @@ import {Grid, Column} from '@twilio-paste/grid';
 import {Select, Option} from '@twilio-paste/select';
 import {useUID} from '@twilio-paste/uid-library';
 
+import {SITE_TOPBAR_HEIGHT} from '../../constants';
+
 export interface TokensListFilterProps {
   value: string;
   handleThemeChange: (e: React.ChangeEvent<HTMLSelectElement>) => void;
@@ -44,7 +46,7 @@ export const TokensListFilter: React.FC<React.PropsWithChildren<TokensListFilter
     <Box
       marginBottom="space80"
       position="sticky"
-      top="0"
+      top={`${SITE_TOPBAR_HEIGHT}px`}
       zIndex="zIndex10"
       marginX="spaceNegative40"
       css={{